Change temporaryWebappPath and webappPath to be same
Making these paths the same means web-dev resources can be modified in
the webapp folder without having to redeploy or mess with symlinks in
order to have a productive environment. Since these paths were changed
compiled classes and libraries need to be ignored by git.

Also added jetty resources so jetty can be executed from within sbt and
it will support the use of JSPs. Changed Jersey dependency to
jersey-server because it contains the servlet needed.