Rewrite UserDao so it is DRYer
save method now takes a type parameter determining which method to
delegate to the new saveUser or saveEmployee methods. Also created
writeProperities methods which create the data on a node for a User or
Employee object. Refactored out the logic for retrieving a node so it
isn't written in each saving method.

Created getUser method to retrieve only the information needed for a
User reference to avoid retrieving the whole User/Employee object graph
when fetching an Employee with get. get method has also been updated to
pull up ContentInfo data.