Create Response objects for business methods
Update UserManager to call validation before registering a user. Update
callers of UserManager.register to expect and handle the new Response
objects. Update UserManagerSpecs by creating and using ResponseMatchers.
Add an Error type to ValidationResult, which isn't really a
ValidationResult but needed to be used with the Failure Response type.