Adjust object model for CaseStudy to reflect wireframes
Wireframes include a video and images which were both missing.
Additionally it is more logical to think of the assets for a case study
as children of the case study in the jcr repository. The downside of
this on save of CaseStudy the uuids of the assets do not get updated
unless the CaseStudy is pulled from the repository again. Due to this
the specs had to change to only compare the uuids of clients and case
studies on retrieval. Other data could be checked as well but it would
take a great deal of code to traverse the object graph.