Create save method to test if user exists
Add a save method to CaseStudyManager which accepts Option[User] and
tests if the user exists. If not signify Failure with a ValidationError.
The addition of this method simplifies the CaseStudyServlet so it
doesn't need to do all the checking for an existing user.

Also add Success and Failure aliases to WkPredef so classes won't have
to explicitly include them from the business package. Update
caseStudy-basic.vm to handle the new information provided by
CaseStudyServlet (sort of, doesn't list errors).