Delete MappingServlet add capabilities to WkServlet
WkServlet takes over for MappingServlet in that it now checks the
init-params of a servlet for matching views. Views in web.xml must
have the format appended to them.

Fix a typo in client.vm causing the name of the client to not show
up.

Adjust ClientServlet to make use of new WkServlet functionality.