Change UserManager.save to not call encryptPassword
Remove decision logic from UserManager.encryptPassword, it is now up to
the callee to determine whether or not the password should be encrypted
and when.

Update various locations calling UserManager.save to first call
UserManager.encryptPassword. These locations are all representative of
creating users for the first time and not editing/updating existing
users.